refactor: refactor leveled compactor structure and move flush method (#441)
* refactor: Added RwLock abstraction to the Executor trait

* refactor: refactor leveled compactor structure

* refactor: move flush logic

* chore: clippy and fmt

* chore: rebase

* chore: pass CI test

* Merge fix

* fix: fix the logic order
This commit block the transaction creation during compaction by adding a large write lock.
After this commit, the logic flow in logs seems to be correct.
It is temporary fix and should be removed.
The only test case involved is `test_read_write_dyn`.
The root cause I think, is from the seperation of flush and remove immutables logic, from sync to async. Callback will still have time interval so it cannot solve this problem.
In logs, there are points shows:
- There are transaction creating during compaction.
- The immutables is modified during transaction.
- The snapshot in transaction sometimes cannot see the latest flushed SST, even if the snapshot ts is greater then the ts after manifest update.
- There exists race condition, and after time-consuming logs, the pass rate will fall, suggest there exist time window. After this commit, the test will return to all pass.

* chore: remove a helper method

* test: add plug-in compactor test
Add a test to test the plug-in compactor.
Change visibility of some mod and struct, especially in `version`, `scope`, and mod in lib.

* fix: pass WASM test

* chore: fix comments and some layout

* fix: fix the minor_compaction sort logic
solves #456

* fix: remove dead code instruction

* refactor: release lock before `check_then_compaction`

* chore: make clippy happy

* fix: compaction task spin lock awareness logic in snapshot and scan

* chore: change LeveledTask visivility and some err code

* Update src/compaction/leveled.rs

* Update src/compaction/leveled.rs

* fix: compiling error

---------

Co-authored-by: Gwo Tzu-Hsing <gotzehsing@gmail.com>
Co-authored-by: 123789456ye <123789456ye@gmail.com>